requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
blinker==1.6.2
cffi==1.15.1
click==8.1.3
coverage==7.2.6
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
hiredis==2.2.3
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
pycparser==2.21
pytest==7.3.1
python-dotenv==1.0.0
redis==4.5.5
SQLAlchemy==2.0.15
typing_extensions==4.6.2
Werkzeug==2.3.4
